Mechanism for rendering razor component from minimal API. Show we can do a JS-free counter.
